+++ frontend/wailsjs/go/app/ComposerApp.d.ts
@@ -24,6 +24,8 @@ export function GetAllAccountIdentities():Promise<Array<app.AccountIdentityGroup
 
 export function GetComposeMode():Promise<app.ComposeMode>;
 
+export function GetDarkComposerBody():Promise<boolean>;
+
 export function GetDraft():Promise<smtp.ComposeMessage>;
 
 export function GetIdentities(arg1:string):Promise<Array<account.Identity>>;